change "rlock" back to non-recursive "lock"
