+++ package.json
@@ -41,30 +41,38 @@
   "scripts": {
     "clean": "rm -rf build coverage .nyc_output browserify-tests node browser dist",
     "esbuild": "rm -rf dist && node ./scripts/build.mjs && yarn build-types",
-    "build-tsc": "tsc",
     "build-types": "dts-bundle-generator --config dts-bundle.config.json",
+
+    "build-tsc": "tsc",
     "clean-build": "yarn clean && yarn build-tsc",
     "clean-build-test": "yarn clean && yarn build-tsc && yarn test",
+
     "print-platform-support": "yarn clean && yarn build-tsc && node build/print-platform-support.js",
     "start": "node build/example-app.js",
+
     "browserify-tests": "mkdir -p browserify-tests && browserify build/test/universal/*.test.js build/test/browser/*.test.js build/test/browser-run-exit.js > browserify-tests/bundle.test.js",
     "browser-run-tests": "cat browserify-tests/bundle.test.js | browser-run | tap-spec",
     "test.browser": "yarn clean && yarn build-tsc && yarn browserify-tests && yarn browser-run-tests",
     "browserify-benchmark": "mkdir -p browserify-tests && browserify build/test/browser/*.benchmark.js > browserify-tests/bundle.benchmark.js",
     "browser-run-benchmark": "cat browserify-tests/bundle.benchmark.js | browser-run",
     "benchmark.browser": "yarn clean && yarn build-tsc && mkdir -p benchmark-results && yarn browserify-benchmark && yarn browser-run-benchmark | tee benchmark-results/benchmark.browser.txt",
+
     "test.node": "yarn test",
     "benchmark.node": "yarn clean && yarn build-tsc && mkdir -p benchmark-results && node --expose-gc build/test/node/*.benchmark.js | tee benchmark-results/benchmark.node.txt",
+
     "test": "tap --timeout=120 --no-bail build/test/universal/*.test.js build/test/node/*.test.js",
     "test-verbose": "tap -Rspecy --no-bail --jobs=1 build/test/universal/*.test.js build/test/node/*.test.js",
     "test-coverage": "yarn test --coverage-report=lcov",
+    "watch-test": "onchange -i src/*.ts src/test/*.ts -- yarn clean-build-test",
+
     "test-crypto": "tap -Rspecy --no-bail --jobs=1 build/test/node/crypto.node.test.js",
-    "test-rlock": "tap -Rspecy --no-bail --jobs=1 build/test/universal/rlock.test.js",
+    "test-lock": "tap -Rspecy --no-bail --jobs=1 build/test/universal/lock.test.js",
     "test-peer-client-server": "tap -Rspecy --no-bail --jobs=1 build/test/node/peer-client-server.node.test.js",
     "test-storage": "tap -Rspecy --no-bail --jobs=1 build/test/node/storage-async.node.test.js",
     "test-storage-driver": "tap -Rspecy --no-bail --jobs=1 build/test/node/storage-driver-async.node.test.js",
-    "watch-test": "onchange -i src/*.ts src/test/*.ts -- yarn clean-build-test",
+
     "prepublishOnly": "yarn clean && yarn esbuild",
+
     "depchart-main": "depchart     `find src | grep .ts` --exclude src/print-platform-support.ts src/decls.d.ts src/index.ts src/index.browser.ts src/shims/*.ts src/entries/*.ts `find src | grep '/test/'` `find src | grep '/util/'` --rankdir LR -o depchart/depchart-main --node_modules omit",
     "depchart-no-types": "depchart `find src | grep .ts` --exclude src/print-platform-support.ts src/decls.d.ts src/index.ts src/index.browser.ts src/shims/*.ts src/entries/*.ts `find src | grep '/test/'` `find src | grep '/util/'` `find src | grep types.ts` --rankdir LR -o depchart/depchart-no-types --node_modules omit",
     "depchart-deps": "depchart     `find src | grep .ts` --exclude src/print-platform-support.ts src/decls.d.ts src/index.ts src/index.browser.ts src/shims/*.ts src/entries/*.ts `find src | grep '/test/'` `find src | grep '/util/'` --rankdir LR -o depchart/depchart-deps --node_modules integrated",
